requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.2
decorator==5.1.1
Django==5.0.1
django-defender==0.9.7
django-filter==23.5
django-lifecycle==1.1.2
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
easy-thumbnails==2.8.5
executing==2.0.1
geographiclib==2.0
geopy==2.4.1
inflection==0.5.1
iniconfig==2.0.0
ipython==8.21.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
phonenumbers==8.13.29
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg==3.1.17
psycopg-binary==3.1.17
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pytest==8.0.0
pytest-django==4.8.0
python-decouple==3.8
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
wcwidth==0.2.13
